// ExistingCode/src/data-access/domain/tables/assaybatchstatuslog/schema.ts
export const AssayBatchStatusLogTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	assayBatchStatusLogId: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	logDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	labCode: z.string().max(20),
	/**
	 * dataType: nvarchar
	 */
	batchNo: z.string().max(50),
	/**
	 * dataType: bit
	 */
	validated: z.any(),
	/**
	 * dataType: nvarchar
	 */
	batchStatus: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	createdBy: z.string().max(50),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});